mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.20.0",
    "langchain-ollama>=0.3.0",
    "mcp>=1.6.0",
    "ollama>=0.4.7",
    "openai>=1.69.0",
    "python-dotenv>=1.1.0",
]
